requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pillow>=10.0.0
reportlab>=4.1.0
chromadb>=0.5.0
rank-bm25>=0.2.2
requests>=2.31.0
pydicom>=2.4.0
numpy>=1.26.0
tiktoken>=0.6.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
crewai>=0.80.0